# Conflicts:
#	pom.xml
#	src/main/java/gg/projecteden/nexus/features/chat/Chat.java
#	src/main/java/gg/projecteden/nexus/features/commands/staff/admin/DatabaseCommand.java
#	src/main/java/gg/projecteden/nexus/features/discord/Bot.java
#	src/main/java/gg/projecteden/nexus/features/discord/Discord.java
#	src/main/java/gg/projecteden/nexus/features/discord/commands/DiscordDiscordCommand.java
#	src/main/java/gg/projecteden/nexus/features/discord/commands/HoursDiscordCommand.java
#	src/main/java/gg/projecteden/nexus/features/discord/commands/NewPlayersDiscordCommand.java
#	src/main/java/gg/projecteden/nexus/features/discord/commands/NicknameDiscordCommand.java
#	src/main/java/gg/projecteden/nexus/features/discord/commands/PronounsDiscordCommand.java
#	src/main/java/gg/projecteden/nexus/features/discord/commands/SubscribeDiscordCommand.java
#	src/main/java/gg/projecteden/nexus/features/discord/commands/SuggestDiscordCommand.java
#	src/main/java/gg/projecteden/nexus/features/discord/commands/TicketsDiscordCommand.java
#	src/main/java/gg/projecteden/nexus/features/discord/commands/TwitterDiscordCommand.java
#	src/main/java/gg/projecteden/nexus/features/discord/commands/UnsubscribeDiscordCommand.java
#	src/main/java/gg/projecteden/nexus/features/discord/commands/justice/AltsDiscordCommand.java
#	src/main/java/gg/projecteden/nexus/features/discord/commands/justice/ClearChatDiscordCommand.java
#	src/main/java/gg/projecteden/nexus/features/discord/commands/justice/HistoryDiscordCommand.java
#	src/main/java/gg/projecteden/nexus/features/discord/commands/justice/_PunishmentDiscordCommand.java
#	src/main/java/gg/projecteden/nexus/features/fakenpc/FakeNPCPacketUtils.java
#	src/main/java/gg/projecteden/nexus/features/minigames/mechanics/Sabotage.java
#	src/main/java/gg/projecteden/nexus/features/resourcepack/ResourcePack.java
#	src/main/java/gg/projecteden/nexus/features/store/HandlePurchaseCommand.java
#	src/main/java/gg/projecteden/nexus/features/store/perks/autosort/AutoSortFeature.java
#	src/main/java/gg/projecteden/nexus/framework/interfaces/PlayerOwnedObject.java

+++ src/main/java/gg/projecteden/nexus/models/boost/Booster.java
@@ -27,7 +27,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
 
-import static gg.projecteden.utils.StringUtils.camelCase;
+import static gg.projecteden.nexus.utils.StringUtils.camelCase;
 
 @Data
 @Entity(value = "booster", noClassnameStored = true)
@@ -50,7 +50,7 @@ public class Booster implements PlayerOwnedObject {
 		private UUID uuid;
 		private Boostable type;
 		private double multiplier;
-		private int duration;
+		private long duration;
 		private LocalDateTime received;
 		private LocalDateTime activated;
 		private boolean cancelled;
@@ -59,7 +59,7 @@ public class Booster implements PlayerOwnedObject {
 			this(uuid, type, multiplier, duration.get() / 20);
 		}
 
-		public Boost(@NonNull UUID uuid, Boostable type, double multiplier, int duration) {
+		public Boost(@NonNull UUID uuid, Boostable type, double multiplier, long duration) {
 			this.uuid = uuid;
 			this.id = getBooster().getBoosts().size();
 			this.type = type;
@@ -174,9 +174,9 @@ public class Booster implements PlayerOwnedObject {
 			return Timespan.of(getExpiration()).format() + " left";
 		}
 
-		public int getDurationLeft() {
+		public long getDurationLeft() {
 			if (isActive())
-				return (int) ChronoUnit.SECONDS.between(LocalDateTime.now(), getExpiration());
+				return ChronoUnit.SECONDS.between(LocalDateTime.now(), getExpiration());
 			else
 				return duration;
 		}
@@ -205,7 +205,7 @@ public class Booster implements PlayerOwnedObject {
 		return boost;
 	}
 
-	public Boost add(Boostable type, double multiplier, int duration) {
+	public Boost add(Boostable type, double multiplier, long duration) {
 		Boost boost = new Boost(uuid, type, multiplier, duration);
 		add(boost);
 		return boost;
